requirements.txt
paho-mqtt
pymongo
python-dotenv
sqlalchemy
mysql-connector-python